Refactor Builtin<->VHLO type converter into its own class
